+++ packages/sandbox_next/pages/index.tsx
@@ -8,14 +8,14 @@ export default Demo;
 
 // Get current `jsonTree` and `zipConfig` from session
 // If `jsonTree` is missing, will be loaded from `data` dir
-// If `zipConfig` is missing, will be created from `lib/config` (based on `CoreConfig`) and compressed with `compressConfig`
+// If `zipConfig` is missing, will be created from `lib/config_ui` (based on `CoreConfig`) and compressed with `compressConfig`
 export const getServerSideProps = withSessionSsr<DemoQueryBuilderProps>(
   async function getServerSideProps({ req }) {
     const sessionData = await getSessionData(req);
     return {
       props: {
         jsonTree: sessionData?.jsonTree || getInitialTree(),
-        zipConfig: sessionData?.zipConfig || getInitialConfig(),
+        zipConfig: sessionData?.zipConfig || getInitialZipConfig(),
       }
     };
   }