+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
+- 6.2.0
+  - Allow saving and loading config from server (PR #866) (issue #817)
+    - New utils in `Utils.ConfigUtils`: `compressConfig()`, `decompressConfig()`
+    - New settings option `useConfigCompress`
+    - Added new test app `sandbox_next` to demonstrate new server-side features
+  - Export utils (like `mongoFormatOp1`, `mongoFormatOp2`) in `CoreConfig.ctx.utils` (PR #866) (issue #890)
 - 6.1.2
   - Fix typings for `children1`: tuple -> array (PR #885) (issue #881)
   - Fix compare fields for case when type != widget.type (PR #875) (issue #758)